requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
openai>=1.0.0
python-docx>=1.0.0
pdfplumber>=0.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0

